Replace bitwise operators with Q_DECLARE_OPERATORS_FOR_FLAGS
Use default QFlags constructor instead of the deprecated constructor
that takes nullptr.

Remove static_cast<int> inside QCOMPARE in
TestDUChain::testReparseOnDocumentActivated(), because Qt Test prints
QFlags as integer Actual and Expected values.

GIT_SILENT

+++ plugins/clang/tests/test_duchain.cpp
@@ -1080,8 +1080,8 @@ void TestDUChain::testReparseOnDocumentActivated()
     {
         DUChainReadLocker lock;
         auto ctx = file.topContext();
-        QCOMPARE(static_cast<int>(ctx->features() & TopDUContext::AllDeclarationsContextsAndUses),
-                 static_cast<int>(TopDUContext::AllDeclarationsContextsAndUses));
+        QCOMPARE(ctx->features() & TopDUContext::AllDeclarationsContextsAndUses,
+                 TopDUContext::AllDeclarationsContextsAndUses);
         QVERIFY(ctx->topContext()->ast());
     }
 }
